FEAT: Códigos de gerenciamento de estado aplicados, porém não funcionando ainda. #65
