+++ src/main.js
@@ -1,10 +1,14 @@
 import './assets/main.css'
 
 import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+
+import router from './router'
 import App from './App.vue'
-import router from './router';
 
-import '@fortawesome/fontawesome-free/css/all.css'
-import '@fortawesome/fontawesome-free/js/all.js'
+const pinia = createPinia()
 
-createApp(App).use(router).mount('#app')
+const app = createApp(App)
+app.use(pinia)
+app.use(router)
+app.mount('#app')
