+++ src/stores/book.js
@@ -1,0 +1,66 @@
+import { ref } from 'vue';
+import { defineStore } from 'pinia';
+
+export const useBooksStore = defineStore('books', () => {
+  const books = ref([
+    const produtos = [
+  {
+    id: 1,
+    titulo: 'Chain of Iron: Volume 2',
+    autor: 'Cassandra Clare',
+    preco: 23.24,
+    capa: '/src/chain_of_iron.png',
+  },
+  {
+    id: 2,
+    titulo: 'Chain of Thorns',
+    autor: 'Cassandra Clare',
+    preco: 23.24,
+    capa: '/src/chain_of_thorns.png',
+  },
+  {
+    id: 3,
+    titulo: 'City of Fallen Angels',
+    autor: 'Cassandra Clare',
+    preco: 13.94,
+    capa: '/src/city_of_fallen_angels.png',
+  },
+  {
+    id: 4,
+    titulo: 'Nona the Ninth',
+    autor: 'Cassandra Clare',
+    preco: 16.84,
+    capa: '/src/nona_the_niinth.png',
+  },
+  {
+    id: 5,
+    titulo: 'Harlem Shuffle',
+    autor: 'Colson Whitehead',
+    preco: 26.92,
+    capa: '/src/harlem_shuffle.png',
+  },
+  {
+    id: 6,
+    titulo: 'Two Old Women',
+    autor: 'Velma Wallis',
+    preco: 13.95,
+    capa: '/src/two_old_women.png',
+  },
+  {
+    id: 7,
+    titulo: 'Carrie Soto Is Back',
+    autor: 'Taylor Jenkins Reid',
+    preco: 26.04,
+    capa: '/src/carrie_soto_is_back.png',
+  },
+  {
+    id: 8,
+    titulo: 'Book Lovers',
+    autor: 'Emily Henry',
+    preco: 15.81,
+    capa: '/src/book_lovers.png',
+  },
+];
+
+return { books };
+})
